refactor(backend): extract settings document routes into app/routes/settings.py (#65)
Move GET/POST /internal/settings-document to APIRouter.
main.py: 2355 -> 2328 lines. 26 routes remain inline.
